Revert "Feat/router (#1)"
This reverts commit 33f5bc79cb4d146e1afe52840806c58029c1f62f.
