+++ .gitignore
@@ -1,4 +1,3 @@
 /node_modules
-/request
 .env
-.DS_Store
+/request
