+++ package.json
@@ -11,8 +11,8 @@
   },
   "type": "module",
   "scripts": {
-    "dev": "nodemon src/app.js",
-    "start": "node src/app.js"
+    "dev": "nodemon app.js",
+    "start": "node app.js"
   },
   "prisma": {
     "seed": "node prisma/seed.js"
