--- src/async-handler.js
@@ -1,9 +1,0 @@
-export default function asyncHandler(handler) {
-  return async function (req, res, next) {
-    try {
-      await handler(req, res);
-    } catch (e) {
-      next(e);
-    }
-  };
-}
